Bug fix: Types

+++ main/src/Tools.cpp
@@ -117,28 +117,28 @@ void read_muon_events_into_arrays( Muon::HitsSoA *muon_station_hits,
       std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_tile[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_x[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_x[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dx[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dx[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_y[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_y[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dy[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dy[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_z[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_z[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dz[station_offset]) );
+      std::copy_n((float*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_dz[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
       std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_uncrossed[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
-      std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_time[station_offset]) );
+      std::copy_n((unsigned int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_time[station_offset]) );
       raw_input += sizeof(int) * number_of_hits;
 
       std::copy_n((int*) raw_input, number_of_hits, &( muon_station_hits[i_event].m_delta_time[station_offset]) );
